// src/modules/Companies/utils/types.ts
export interface Company {
  id?: string;
  name: string;
  razon_social_c: string;
  direccion_c: string;
  resolucion_ministerial_c: string;
  identificacion_fiscal_c?:string;
  modified_user_id?:string;
  created_by?:string;
  //idgrupocliente_c: string;
}